+++ docs/changes.rst
@@ -9,17 +9,23 @@ skops Changelog
     :depth: 1
     :local:
 
+v0.6
+----
+
 v0.5
 ----
 - Added CLI entrypoint support (:func:`.cli.entrypoint.main_cli`)
   and a command line function to convert Pickle files
   to Skops files (:func:`.cli._convert.main`). :pr:`249` by `Erin Aho`_
 - Support more array-like data types for tabular data and list-like data types
-  for text data. :pr:`179` by `Francesco Cariaggi`_.
+  for text data. :pr:`179` by :user:`Francesco Cariaggi <anferico>`.
 - Add an option `use_intelex` to :func:`skops.hub_utils.init` which, when
   enabled, will result in the Hugging Face inference API running with Intel's
   scikit-learn intelex library, which can accelerate inference times. :pr:`267`
   by `Benjamin Bossan`_.
+- Model cards that have been written into a markdown file can now be parsed back
+  into a :class:`skops.card.Card` object and edited further by using the
+  :func:`skops.card.parse_modelcard` function. :pr:`257` by `Benjamin Bossan`_.
 
 v0.4
 ----